2 bug fixes in master branch
In pytorch_run.py: now using target_dir == -1 instead of target_dir is -1

In external/vgg.py: I now allow passing a device parameter in the __init__
call.  Otherwise it starts by loading a model in cuda:0. This fails if cuda:0
is already busy on a different tas and runs out of memory.